+++ chance.lua
@@ -48,6 +48,17 @@ chance.VERSION = setmetatable(
         end
 })
 
+--- Make a shallow copy of a table.
+--
+-- @local
+local function makeShallowCopy(array)
+    local copy = {}
+    for _,value in ipairs(array) do
+        table.insert(copy, value)
+    end
+    return copy
+end
+
 --- Seeds the random number generator.
 --
 -- This function accepts one parameter: a seed, which it uses to seed
@@ -104,6 +115,90 @@ function chance.random(m, n)
         end
     end
     return math.random()
+end
+
+--- Sets of data which some functions choose from.
+--
+-- Many functions select random data from a predefined source.  For
+-- example, @{chance.month} randomly picks a name from an existing
+-- list of names.  This table contains all of those types of
+-- predefined sets of data.  Developers can modify or add new sets
+-- of data by using the @{chance.set} function.
+--
+-- The keys for this table must strings, which name the data set.
+--
+-- The values must either be arrays (which can contain any types of
+-- values), or a single function.  If the value is a function then the
+-- library treats it as a generator for that data set, i.e. the
+-- library will invoke that function expecting it to return the
+-- appropriate type of random data.  The function will receive no
+-- arguments.
+--
+-- @see chance.set
+-- @see chance.fromSet
+-- @field dataSets
+chance.dataSets = {}
+
+--- Define or modify a set of data.
+--
+-- This function creates a new set of data or replaces an existing
+-- one.  The key parameter must be a string naming the data set.  The
+-- data parameter must be either a table of data, which can be of any
+-- type, or must be a function.  If it is a function then the library
+-- treats it as a generator and will invoke that function with no
+-- arguments whenever random data is requested from that set.
+--
+-- @see chance.fromSet
+-- @see chance.dataSets
+--
+-- @tparam string key
+-- @tparam table|function data
+-- @treturn nil
+function chance.set(key, data)
+    chance.dataSets[key] = data
+end
+
+--- Add data to an existing data set.
+--
+-- See the documentation on @{chance.set} for details on the
+-- <code>key</code> parameter.  The <code>data</code> must be a table
+-- of values which the function will add to the existing data set.
+-- <strong>This does not work for data sets that have generator
+-- functions for their values.</strong>
+--
+-- @see chance.set
+-- @see chance.dataSets
+--
+-- @tparam string key
+-- @tparam table data
+-- @treturn nil
+function chance.appendSet(key, data)
+    for _,value in ipairs(data) do
+        table.insert(chance.dataSets[key], value)
+    end
+end
+
+--- Select random data from an existing data set.
+--
+-- See the documentation on @{chance.set} for details on the
+-- restrictions and semantics of the <code>key</code> parameter.
+--
+-- @see chance.set
+-- @see chance.dataSets
+--
+-- @tparam string|function key
+-- @return Random data of potentially any type, or nil if there is no
+-- data set for the given <code>key</code>
+function chance.fromSet(key)
+    local data = chance.dataSets[key]
+
+    if data == nil then return nil end
+
+    if type(data) == "function" then
+        return data()
+    else
+        return chance.pick(data)
+    end
 end
 
 --- Basic
@@ -451,11 +546,13 @@ function chance.year(flags)
     return chance.random(minimum, maximum)
 end
 
---- Returns the name of a random month.
+--- Names of months.
 --
--- @treturn string
-function chance.month()
-    return chance.pick {
+-- @see chance.month
+-- @local
+-- @field months
+-- @table chance.dataSets
+chance.set("months", {
         "January",
         "February",
         "March",
@@ -467,15 +564,41 @@ function chance.month()
         "September",
         "October",
         "November",
-        "December"
-    }
+        "December",
+})
+
+--- Returns the name of a random month.
+--
+-- This function chooses the name of a month from the
+-- <code>months</code> data set.
+--
+-- @treturn string
+function chance.month()
+    return chance.fromSet("months")
 end
+
+--- Names of days of the week.
+--
+-- @see chance.day
+-- @local
+-- @field days
+-- @table chance.dataSets
+chance.set("days", {
+        "Monday",
+        "Tuesday",
+        "Wednesday",
+        "Thursday",
+        "Friday",
+        "Saturday",
+        "Sunday",
+})
 
 --- Returns a random day of the week.
 --
--- By default this function will return the name of a day of the week.
--- However, it accepts an optional table of flags which control the
--- possible days it returns.  If the flags <code>weekdays</code> or
+-- By default this function will return the name of a day of the week,
+-- chosen from the <code>days</code> data set.  The function accepts
+-- an optional table of flags which control the possible days it
+-- returns.  If the flags <code>weekdays</code> or
 -- <code>weekends</code> are false then the function will not return
 -- those types of days.
 --
@@ -486,15 +609,7 @@ end
 -- @param[opt] flags
 -- @treturn string
 function chance.day(flags)
-    local days = {
-        "Monday",
-        "Tuesday",
-        "Wednesday",
-        "Thursday",
-        "Friday",
-        "Saturday",
-        "Sunday",
-    }
+    local days = makeShallowCopy(chance.dataSets["days"])
 
     -- This logic takes advantage of the specific order of the `days`
     -- table above.
@@ -698,15 +813,6 @@ function chance.pick(array, count)
     end
 
     return array[chance.random(1, size)]
-end
-
---- Make a shallow copy of a table.
-local function makeShallowCopy(array)
-    local copy = {}
-    for _,value in ipairs(array) do
-        table.insert(copy, value)
-    end
-    return copy
 end
 
 --- Randomly shuffle the contents of an array.